Refactor EPUB paragraph retrieval and enhance TypeScript definitions
- Improved the `getAllParagraphsForBook` function by optimizing the retrieval of views and mapping paragraphs to their respective sections.
- Introduced a `batchEmbed` function in `sql.ts` to handle embedding parameters in batches, enhancing performance during processing.
- Updated the `Vector` struct to change the `id` type from `u32` to `u64` for better compatibility with larger datasets.
- Updated timestamps in auto-generated TypeScript files to reflect the latest generation time.

+++ src-tauri/src/vectordb.rs
@@ -12,12 +12,12 @@ use tauri::Manager;
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub struct Vector {
-    pub id: u32,
+    pub id: u64,
     pub vector: Vec<f32>,
 }
 
 impl Vector {
-    pub fn new(id: u32, vector: Vec<f32>) -> Self {
+    pub fn new(id: u64, vector: Vec<f32>) -> Self {
         Self { id, vector }
     }
 }
